staging bugfixing

+++ sample_app/lib/tasks/tmp.rake
@@ -8,6 +8,9 @@ namespace :rafolo do
       Rake::Task["tmp:clear"].invoke
 
       path = File.expand_path('tmp', Rails.root)
+      if !File.exists?(path)
+        return
+      end
       Dir[path + '/**/*.pid'].each {|f|
         FileUtils.rm(f.to_s)
       }
